restructure component models
