+++ experiments/AMIP/components/ocean/slab_ocean.jl
@@ -10,6 +10,9 @@ import ClimaCoupler.BCReader: float_type_bcf
 
 include("../slab_utils.jl")
 
+###
+### Functions required by ClimaCoupler.jl for a SurfaceModelSimulation
+###
 """
     SlabOceanSimulation{P, Y, D, I}
 
@@ -41,18 +44,6 @@ end
 name(::SlabOceanSimulation) = "SlabOceanSimulation"
 
 """
-    temp_anomaly(coord)
-
-Calculates a an anomaly to be added to the initial condition for temperature.
-This default case includes only an anomaly at the tropics.
-"""
-function temp_anomaly(coord)
-    # include tropics anomaly
-    anom = FT(29 * exp(-coord.lat^2 / (2 * 26^2)))
-    return anom
-end
-
-"""
     slab_ocean_space_init(space, params)
 
 Initialize the slab ocean prognostic variable (temperature), including an
@@ -70,15 +61,6 @@ function slab_ocean_space_init(space, params)
     Y = ClimaCore.Fields.FieldVector(T_sfc = T_sfc)
 
     return Y, space
-end
-
-# ode
-function slab_ocean_rhs!(dY, Y, cache, t)
-    p, F_turb_energy, F_radiative, area_fraction = cache
-    FT = eltype(Y.T_sfc)
-    rhs = @. -(F_turb_energy + F_radiative) / (p.h * p.ρ * p.c)
-    @. dY.T_sfc = rhs * Regridder.binary_mask(area_fraction) * p.evolving_switch
-    @. cache.q_sfc = TD.q_vap_saturation_generic.(cache.thermo_params, Y.T_sfc, cache.ρ_sfc, TD.Liquid())
 end
 
 """
@@ -127,40 +109,41 @@ function ocean_init(
     return sim
 end
 
-# file specific
-"""
-    scale_sst(SST::FT, _info)
-Ensures that the space of the SST struct matches that of the land_fraction, and converts the units to Kelvin (N.B.: this is dataset specific)
-"""
-scale_sst(SST, _info) = (swap_space!(zeros(axes(_info.land_fraction)), SST) .+ float_type_bcf(_info)(273.15))
-
 # extensions required by Interfacer
-get_field(sim::SlabOceanSimulation, ::Val{:surface_temperature}) = sim.integrator.u.T_sfc
-get_field(sim::SlabOceanSimulation, ::Val{:surface_humidity}) = sim.integrator.p.q_sfc
-get_field(sim::SlabOceanSimulation, ::Val{:roughness_momentum}) = sim.integrator.p.params.z0m
-get_field(sim::SlabOceanSimulation, ::Val{:roughness_buoyancy}) = sim.integrator.p.params.z0b
-get_field(sim::SlabOceanSimulation, ::Val{:beta}) = convert(eltype(sim.integrator.u), 1.0)
-get_field(sim::SlabOceanSimulation, ::Val{:surface_albedo}) = sim.integrator.p.params.α
-get_field(sim::SlabOceanSimulation, ::Val{:area_fraction}) = sim.integrator.p.area_fraction
 get_field(sim::SlabOceanSimulation, ::Val{:air_density}) = sim.integrator.p.ρ_sfc
+get_field(sim::SlabOceanSimulation, ::Val{:area_fraction}) = sim.integrator.p.area_fraction
+get_field(sim::SlabOceanSimulation, ::Val{:beta}) = convert(eltype(sim.integrator.u), 1.0)
+get_field(sim::SlabOceanSimulation, ::Val{:roughness_buoyancy}) = sim.integrator.p.params.z0b
+get_field(sim::SlabOceanSimulation, ::Val{:roughness_momentum}) = sim.integrator.p.params.z0m
+get_field(sim::SlabOceanSimulation, ::Val{:surface_albedo}) = sim.integrator.p.params.α
+get_field(sim::SlabOceanSimulation, ::Val{:surface_humidity}) = sim.integrator.p.q_sfc
+get_field(sim::SlabOceanSimulation, ::Val{:surface_temperature}) = sim.integrator.u.T_sfc
+get_field(sim::SlabOceanSimulation, ::Val{:water}) = nothing
+
+"""
+    get_field(sim::SlabOceanSimulation, ::Val{:energy})
+
+Extension of Interfacer.get_field to get the energy of the ocean.
+It multiplies the the slab temperature by the heat capacity, density, and depth.
+"""
+get_field(sim::SlabOceanSimulation, ::Val{:energy}) =
+    sim.integrator.p.params.ρ .* sim.integrator.p.params.c .* sim.integrator.u.T_sfc .* sim.integrator.p.params.h
 
 function update_field!(sim::SlabOceanSimulation, ::Val{:area_fraction}, field::ClimaCore.Fields.Field)
     sim.integrator.p.area_fraction .= field
 end
-
-function update_field!(sim::SlabOceanSimulation, ::Val{:turbulent_energy_flux}, field)
-    parent(sim.integrator.p.F_turb_energy) .= parent(field)
+function update_field!(sim::SlabOceanSimulation, ::Val{:air_density}, field)
+    parent(sim.integrator.p.ρ_sfc) .= parent(field)
 end
 function update_field!(sim::SlabOceanSimulation, ::Val{:radiative_energy_flux_sfc}, field)
     parent(sim.integrator.p.F_radiative) .= parent(field)
 end
-function update_field!(sim::SlabOceanSimulation, ::Val{:air_density}, field)
-    parent(sim.integrator.p.ρ_sfc) .= parent(field)
+function update_field!(sim::SlabOceanSimulation, ::Val{:turbulent_energy_flux}, field)
+    parent(sim.integrator.p.F_turb_energy) .= parent(field)
 end
 
 # extensions required by FieldExchanger
 step!(sim::SlabOceanSimulation, t) = step!(sim.integrator, t - sim.integrator.t, true)
-
 reinit!(sim::SlabOceanSimulation) = reinit!(sim.integrator)
 
 # extensions required by FluxCalculator (partitioned fluxes)
@@ -182,16 +165,35 @@ function get_model_prog_state(sim::SlabOceanSimulation)
     return sim.integrator.u
 end
 
+###
+### Slab ocean model-specific functions (not explicitly required by ClimaCoupler.jl)
+###
 """
-    get_field(sim::SlabOceanSimulation, ::Val{:energy})
-
-Extension of Interfacer.get_field to get the energy of the ocean.
-It multiplies the the slab temperature by the heat capacity, density, and depth.
+    scale_sst(SST::FT, _info)
+Ensures that the space of the SST struct matches that of the land_fraction, and converts the units to Kelvin (N.B.: this is dataset specific)
 """
-get_field(sim::SlabOceanSimulation, ::Val{:energy}) =
-    sim.integrator.p.params.ρ .* sim.integrator.p.params.c .* sim.integrator.u.T_sfc .* sim.integrator.p.params.h
+scale_sst(SST, _info) = (swap_space!(zeros(axes(_info.land_fraction)), SST) .+ float_type_bcf(_info)(273.15))
 
-get_field(sim::SlabOceanSimulation, ::Val{:water}) = nothing
+# ode
+function slab_ocean_rhs!(dY, Y, cache, t)
+    p, F_turb_energy, F_radiative, area_fraction = cache
+    FT = eltype(Y.T_sfc)
+    rhs = @. -(F_turb_energy + F_radiative) / (p.h * p.ρ * p.c)
+    @. dY.T_sfc = rhs * Regridder.binary_mask(area_fraction) * p.evolving_switch
+    @. cache.q_sfc = TD.q_vap_saturation_generic.(cache.thermo_params, Y.T_sfc, cache.ρ_sfc, TD.Liquid())
+end
+
+"""
+    temp_anomaly(coord)
+
+Calculates a an anomaly to be added to the initial condition for temperature.
+This default case includes only an anomaly at the tropics.
+"""
+function temp_anomaly(coord)
+    # include tropics anomaly
+    anom = FT(29 * exp(-coord.lat^2 / (2 * 26^2)))
+    return anom
+end
 
 """
     dss_state!(sim::SlabOceanSimulation)
@@ -207,4 +209,5 @@ function dss_state!(sim::SlabOceanSimulation)
         buffer = get_dss_buffer(axes(field), p)
         ClimaCore.Spaces.weighted_dss!(field, buffer)
     end
+
 end
