+++ test/component_model_tests/eisenman_seaice_tests.jl
@@ -12,7 +12,7 @@ import CLIMAParameters as CP
 import Thermodynamics as TD
 import Thermodynamics.Parameters as TP
 
-include("../../experiments/AMIP/components/ocean/eisenman_seaice_init.jl")
+include("../../experiments/AMIP/components/ocean/eisenman_seaice.jl")
 
 
 for FT in (Float32, Float64)
